requirements.txt
PyPDF2==3.0.1
pytesseract==0.3.10
pillow==10.3.0
opencv-python-headless==4.8.1.78
opencv-contrib-python==4.8.1.78